+++ openfisca_core/holders.py
@@ -121,7 +121,7 @@ class Holder(object):
         # Check that the requested period matches definition_period
         if column.definition_period != ETERNITY:
             if column.definition_period == MONTH and period.unit != periods.MONTH:
-                raise ValueError('Unable to compute variable {} for period {} : {} must be computed for a whole month. You can use the option DIVIDE or use period.this_month'.format(
+                raise ValueError('Unable to compute variable {} for period {} : {} must be computed for a whole month. You can use the option DIVIDE or use period.first_month'.format(
                     column.name,
                     period,
                     column.name))